1-array/pyproject.toml
[project]
name = "1-array"
version = "0.1.0"
description = ""
requires-python = ">=3.10"
dependencies = [
    "flake8>=7.3.0",
    "matplotlib>=3.10.8",
    "numpy>=2.2.6",
    "pillow>=12.2.0",
    "pip>=26.0.1",
    "pooch>=1.9.0",
    "scipy>=1.15.3",
]
